% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eval_pedigree_ll}
\alias{eval_pedigree_ll}
\alias{eval_pedigree_grad}
\title{Approximate the Log Marginal Likelihood}
\usage{
eval_pedigree_ll(
  ptr,
  par,
  maxvls,
  abs_eps,
  rel_eps,
  indices = NULL,
  minvls = -1L,
  do_reorder = TRUE,
  use_aprx = FALSE,
  n_threads = 1L
)

eval_pedigree_grad(
  ptr,
  par,
  maxvls,
  abs_eps,
  rel_eps,
  indices = NULL,
  minvls = -1L,
  do_reorder = TRUE,
  use_aprx = FALSE,
  n_threads = 1L
)
}
\arguments{
\item{ptr}{object from \code{\link{get_pedigree_ll_terms}}.}

\item{par}{numeric vector with fixed effect coefficients and log scale
parameters.}

\item{maxvls}{maximum number of samples in the approximation.}

\item{abs_eps}{absolute convergence threshold.}

\item{rel_eps}{relative convergence threshold.}

\item{indices}{zero-based vector with indices of which log marginal
likelihood terms to include. Use \code{NULL} if all indices should be
used.}

\item{minvls}{minimum number of samples. Negative values provides a
default which depends on the dimension of the integration.}

\item{do_reorder}{\code{TRUE} if a heuristic variable reordering should
be used. \code{TRUE} is likely the best value.}

\item{use_aprx}{\code{TRUE} if a less precise approximation of
\code{\link{pnorm}} and \code{\link{qnorm}} should be used. This may
reduce the computation time while not affecting the result much.}

\item{n_threads}{number of threads to use.}
}
\value{
\code{eval_pedigree_ll}:
a scalar with the log marginal likelihood approximation.

\code{eval_pedigree}: a vector with the derivatives with
respect to \code{par}. An attribute called \code{"logLik"} contains the
log marginal likelihood approximation.
}
\description{
Approximate the log marginal likelihood and the derivatives with
respect to the model parameters.
}
