% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-partitioning.R
\name{get_biconnected_components}
\alias{get_biconnected_components}
\alias{get_biconnected_components_pedigree}
\title{Finds the Biconnected Components Given}
\usage{
get_biconnected_components(from, to)

get_biconnected_components_pedigree(id, father.id, mother.id)
}
\description{
Finds the Biconnected Components Given
}
