% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-partitioning.R
\name{get_block_cut_tree}
\alias{get_block_cut_tree}
\alias{get_block_cut_tree_pedigree}
\title{Creates a Block Cut Tree Like Object}
\usage{
get_block_cut_tree(from, to)

get_block_cut_tree_pedigree(id, father.id, mother.id)
}
\description{
Creates a Block Cut Tree Like Object
}
