% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-partitioning.R
\name{get_max_balanced_partition}
\alias{get_max_balanced_partition}
\alias{get_max_balanced_partition_pedigree}
\title{Finds an Approximately Balanced Connected Partition}
\usage{
get_max_balanced_partition(
  from,
  to,
  weight_data = NULL,
  slack = 0,
  max_kl_it_inner = 50L,
  max_kl_it = 10000L,
  trace = 0L
)

get_max_balanced_partition_pedigree(
  id,
  father.id,
  mother.id,
  id_weight = NULL,
  slack = 0,
  max_kl_it_inner = 50L,
  max_kl_it = 10000L,
  trace = 0L
)
}
\description{
Uses the method suggested by Chlebíková (1996) to construct an approximate
maximally balanced connected partition. A furhter refinement step can be made
to reduce the cost of the cut edges.
}
