% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pedigree_ll_terms}
\alias{get_pedigree_ll_terms}
\title{Get C++ Object for Log Marginal Likelihood Approximations}
\usage{
get_pedigree_ll_terms(data, max_threads)
}
\arguments{
\item{data}{\code{\link{list}} where each element is a list for a cluster
with an:
\itemize{
  \item{\code{"X"}}{ element with the design matrix for the fixed effect,}
  \item{\code{"y"}}{ element with the zero-one outcomes, and}
  \item{\code{"scale_mats"}}{ element with a list where each element is a
scale/correlation matrix for a particular type of effect.}
}}

\item{max_threads}{maximum number of threads to use.}
}
\description{
Constructs an object needed for \code{\link{eval_pedigree_ll}} and
\code{\link{eval_pedigree_grad}}.
}
