% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvndst}
\alias{mvndst}
\title{Multivariate Normal Distribution CDF}
\usage{
mvndst(
  lower,
  upper,
  mu,
  sigma,
  maxvls = 25000L,
  abs_eps = 0.001,
  rel_eps = 0L,
  minvls = -1L,
  do_reorder = TRUE,
  use_aprx = FALSE
)
}
\arguments{
\item{lower}{numeric vector with lower bounds.}

\item{upper}{numeric vector with upper bounds.}

\item{mu}{numeric vector with means.}

\item{sigma}{covariance matrix.}

\item{maxvls}{maximum number of samples in the approximation.}

\item{abs_eps}{absolute convergence threshold.}

\item{rel_eps}{relative convergence threshold.}

\item{minvls}{minimum number of samples. Negative values provides a
default which depends on the dimension of the integration.}

\item{do_reorder}{\code{TRUE} if a heuristic variable reordering should
be used. \code{TRUE} is likely the best value.}

\item{use_aprx}{\code{TRUE} if a less precise approximation of
\code{\link{pnorm}} and \code{\link{qnorm}} should be used. This may
reduce the computation time while not affecting the result much.}
}
\description{
Provides an approximation of the multivariate normal distribution CDF
over a hyperrectangle.
}
