% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{pedmod_profile}
\alias{pedmod_profile}
\title{Compute Profile Likelihood Based Confidence Intervals}
\usage{
pedmod_profile(
  ptr,
  par,
  delta,
  maxvls,
  minvls = -1L,
  alpha = 0.05,
  abs_eps,
  rel_eps,
  which_prof,
  indices = NULL,
  do_reorder = TRUE,
  use_aprx = FALSE,
  n_threads = 1L,
  cluster_weights = NULL,
  method = 0L,
  seed = 1L,
  verbose = FALSE,
  max_step = 15L,
  standardized = FALSE,
  ...
)
}
\arguments{
\item{ptr}{object from \code{\link{get_pedigree_ll_terms}}.}

\item{par}{numeric vector with the maximum likelihood estimator e.g. from
\code{\link{pedmod_opt}}.}

\item{delta}{numeric scalar with an initial step to take. Subsequent steps
are taken by \code{2^(<iteration number> - 1) * delta}. Two times the
standard error is a good value or a guess thereof. Hessian approximations are
not implemented as of this writing and therefore the user needs to provide
some guess.}

\item{maxvls}{maximum number of samples in the approximation.}

\item{minvls}{minimum number of samples. Negative values provides a
default which depends on the dimension of the integration.}

\item{alpha}{numeric scalar with the confidence level required.}

\item{abs_eps}{absolute convergence threshold for
\code{\link{eval_pedigree_ll}} and \code{\link{eval_pedigree_grad}}.}

\item{rel_eps}{rel_eps convergence threshold for
\code{\link{eval_pedigree_ll}} and \code{\link{eval_pedigree_grad}}.}

\item{which_prof}{integer scalar with index of the parameter which the
profile likelihood curve should be computed for.}

\item{indices}{zero-based vector with indices of which log marginal
likelihood terms to include. Use \code{NULL} if all indices should be
used.}

\item{do_reorder}{\code{TRUE} if a heuristic variable reordering should
be used. \code{TRUE} is likely the best value.}

\item{use_aprx}{\code{TRUE} if a less precise approximation of
\code{\link{pnorm}} and \code{\link{qnorm}} should be used. This may
reduce the computation time while not affecting the result much.}

\item{n_threads}{number of threads to use.}

\item{cluster_weights}{numeric vector with weights for each cluster. Use
\code{NULL} if all clusters have weight one.}

\item{method}{integer with the method to use. Zero yields randomized Korobov
lattice rules while one yields scrambled Sobol sequences.}

\item{seed}{seed to pass to \code{\link{set.seed}} before each gradient and
function evaluation. Use \code{NULL} if the seed should not be fixed.}

\item{verbose}{logical for whether output should be printed to the console
during the estimation of the profile likelihood curve.}

\item{max_step}{integer scalar with the maximum number of steps to take in
either directions.}

\item{standardized}{logical for whether to use the standardized or direct
parameterization. See \code{\link{standardized_to_direct}} and the vignette
at \code{vignette("pedmod", package = "pedmod")}.}

\item{...}{arguments passed on to \code{\link{pedmod_opt}}.}
}
\value{
A list with the following elements
  \item{confs}{2D numeric vector with the profile likelihood based confidence
               interval.}
  \item{xs}{the points at which the profile likelihood is evaluated.}
  \item{p_log_Lik}{the log profile likelihood values at \code{xs}.}
  \item{data}{list with the returned objects from \code{\link{pedmod_opt}}.}
}
\description{
Compute Profile Likelihood Based Confidence Intervals
}
\seealso{
\code{\link{pedmod_opt}} and \code{\link{pedmod_sqn}}.
}
