% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{pedmod_profile_prop}
\alias{pedmod_profile_prop}
\title{Compute Profile Likelihood Based Confidence Intervals for the Proportion
of Variance}
\usage{
pedmod_profile_prop(
  ptr,
  par,
  maxvls,
  minvls = -1L,
  alpha = 0.05,
  abs_eps,
  rel_eps,
  which_prof,
  indices = NULL,
  maxvls_start = max(100L, as.integer(ceiling(maxvls/5))),
  minvls_start = if (minvls < 0) minvls else minvls/5,
  do_reorder = TRUE,
  use_aprx = FALSE,
  n_threads = 1L,
  cluster_weights = NULL,
  method = 0L,
  seed = 1L,
  verbose = FALSE,
  max_step = 15L,
  opt_func = NULL,
  ...
)
}
\arguments{
\item{ptr}{object from \code{\link{pedigree_ll_terms}}.}

\item{par}{numeric vector with the maximum likelihood estimator e.g. from
\code{\link{pedmod_opt}}.}

\item{maxvls}{maximum number of samples in the approximation for each
marginal likelihood term.}

\item{minvls}{minimum number of samples for each
marginal likelihood term. Negative values provides a
default which depends on the dimension of the integration.}

\item{alpha}{numeric scalar with the confidence level required.}

\item{abs_eps}{absolute convergence threshold for
\code{\link{eval_pedigree_ll}} and \code{\link{eval_pedigree_grad}}.}

\item{rel_eps}{rel_eps convergence threshold for
\code{\link{eval_pedigree_ll}} and \code{\link{eval_pedigree_grad}}.}

\item{which_prof}{the index of the random effect which proportion of
variance should be profiled.}

\item{indices}{zero-based vector with indices of which log marginal
likelihood terms to include. Use \code{NULL} if all indices should be
used.}

\item{maxvls_start}{number of samples to use when finding the
initial values for the optimization.}

\item{minvls_start}{number of samples to use when finding the
initial values for the optimization.}

\item{do_reorder}{\code{TRUE} if a heuristic variable reordering should
be used. \code{TRUE} is likely the best value.}

\item{use_aprx}{\code{TRUE} if a less precise approximation of
\code{\link{pnorm}} and \code{\link{qnorm}} should be used. This may
reduce the computation time while not affecting the result much.}

\item{n_threads}{number of threads to use.}

\item{cluster_weights}{numeric vector with weights for each cluster. Use
\code{NULL} if all clusters have weight one.}

\item{method}{integer with the method to use. Zero yields randomized Korobov
lattice rules while one yields scrambled Sobol sequences.}

\item{seed}{seed to pass to \code{\link{set.seed}} before each gradient and
function evaluation. Use \code{NULL} if the seed should not be fixed.}

\item{verbose}{logical for whether output should be printed to the console
during the estimation of the profile likelihood curve.}

\item{max_step}{integer scalar with the maximum number of steps to take in
either directions.}

\item{opt_func}{function to perform minimization with arguments like
\code{\link{optim}}. BFGS is used with \code{\link{optim}} if this argument
is \code{NULL}.}

\item{...}{arguments passed to \code{opt_func}.}
}
\value{
A list like \code{\link{pedmod_profile}}.
}
\description{
Compute Profile Likelihood Based Confidence Intervals for the Proportion
of Variance
}
\details{
The function is only useful when there is more than one type of random
effect. If not, then \code{\link{pedmod_profile}} can be used because of
the scale invariance of the likelihood ratio.
}
\seealso{
\code{\link{pedmod_opt}}, \code{\link{pedmod_sqn}}, and
\code{\link{pedmod_profile}}.
}
